requirements.txt
Adafruit-Blinka==7.0.2
adafruit-circuitpython-dht==3.7.1
adafruit-io==2.6.0
Adafruit-PlatformDetect==3.20.1
Adafruit-PureIO==1.1.9
alabaster==0.7.12
appdirs==1.4.4
astor==0.8.1
attrs==21.4.0
autoflake==1.4
Babel==2.9.1
bandit==1.7.3
black==20.8b1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
coverage==5.5
darglint==1.8.1
distlib==0.3.4
doc8==0.8.1
docutils==0.17.1
eradicate==1.0
filelock==3.6.0
flake8==3.9.2
flake8-2020==1.6.1
flake8-bandit==2.1.2
flake8-broken-line==0.2.1
flake8-bugbear==19.8.0
flake8-builtins==1.5.3
flake8-commas==2.1.0
flake8-comprehensions==3.8.0
flake8-debugger==3.2.1
flake8-docstrings==1.6.0
flake8-eradicate==0.4.0
flake8-fixme==1.1.1
flake8-isort==3.0.1
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-polyfill==1.0.2
flake8-quotes==2.1.2
flake8-rst-docstrings==0.0.12
flake8-string-format==0.2.3
flake8-variables-names==0.0.3
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.2
iniconfig==1.1.1
isort==4.3.21
Jinja2==3.0.3
MarkupSafe==2.1.0
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
paho-mqtt==1.6.1
pathspec==0.9.0
pbr==5.8.1
pep8-naming==0.9.1
Pillow==9.0.1
pkg-resources==0.0.0
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pydocstyle==5.1.1
pyflakes==2.4.0
pyftdi==0.53.3
Pygments==2.11.2
pyparsing==3.0.7
pyserial==3.5
pytest==6.2.5
pytest-cov==2.12.1
python-dev-tools==2020.9.10
pytz==2021.3
pyupgrade==2.31.0
pyusb==1.2.1
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
restructuredtext-lint==1.4.0
rpi-ws281x==4.3.2
RPi.GPIO==0.7.1
RTIMULib==7.2.1
six==1.16.0
smbus==1.1.post2
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==2.4.5
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stevedore==3.5.0
sysv-ipc==1.1.0
testfixtures==6.18.4
tokenize-rt==4.2.1
toml==0.10.2
tox==3.24.5
tox-travis==0.12
typed-ast==1.5.2
typing-extensions==4.1.1
urllib3==1.26.8
virtualenv==20.13.2
wemake-python-styleguide==0.14.1
zipp==3.7.0
